# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JornadaMS API
  version: 1.0.0-mvp1
  description: API REST privada do MVP-1 do JornadaMS.
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
  /api/v1/companies:
    get:
      tags: [Organization]
      operationId: listCompanies
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Organization]
      operationId: createCompany
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
      responses:
        '201':
          description: Empresa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/branches:
    get:
      tags: [Organization]
      operationId: listBranches
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista de filiais
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Organization]
      operationId: createBranch
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchCreate'
      responses:
        '201':
          description: Filial criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sessão iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: []
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Sessão renovada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      x-mvp: MVP-1
      responses:
        '204':
          description: Sessão revogada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      x-mvp: MVP-1
      responses:
        '200':
          description: Contexto do usuário autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: jornada_access
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Conflito de estado ou idempotência
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Dados semanticamente inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Limite de tentativas ou requisições excedido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          minLength: 16
          maxLength: 512
    LoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          description: Segundos até a expiração do access token.
    UserContext:
      type: object
      required: [user_id, email, roles]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        employee_id:
          type: [string, 'null']
          format: uuid
    CompanyCreate:
      type: object
      required: [name, cnpj, default_timezone]
      properties:
        name:
          type: string
          minLength: 1
        cnpj:
          type: string
          minLength: 14
          maxLength: 14
        default_timezone:
          type: string
          minLength: 1
    Company:
      allOf:
        - $ref: '#/components/schemas/CompanyCreate'
        - type: object
          required: [id, status]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [ACTIVE, INACTIVE]
    BranchCreate:
      type: object
      required: [company_id, name, code, timezone]
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
    Branch:
      allOf:
        - $ref: '#/components/schemas/BranchCreate'
        - type: object
          required: [id, status]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [ACTIVE, INACTIVE]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
...
    CompanyPage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Company'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    BranchPage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Branch'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...